Initialize code base
This is original development package from sunplus development kit.
The code quality is bad:
1. source file hierarchy is confusing.
2. naming convention is not confortable.
3. there is no description in the head of source file.
4. there is no description for function definition in source file.

My new code will use unix style. In the future, I will re-orgnize
the directory hierarchy but not now since I don't have time. The
first thing I need to complete is to finish the prototype. My plan
is to make all the improvements when the project is almost done
with much time left.

Good luck to me.