Improve the LCM driver and add ui module as a higher layer.
Considering the tricky ST7587 + GWMS9708D, I set the scan direction
to be: X is normal and Y is reversed.

The scan direction highly depands on the character's modulo. Currently
I use horizonal modulo with ascending order, which means left dot uses
the MSB bit in a byte.

Song font is recommended. Number 9 is used to say hello.

A new UI layer is added into system. All ui stuff will be implemented
in this module.