Now we can use the font from General Plus!
General Plus gives me a font for mono lcd which is vertical negative
with vertical move. So I add a function to transform the dot matrix
data to horizontal positive with horizontal move for my mono lcd.

The driver for font ic GT20 is code completed but it does not work
as expected at this moment. Anyway, it does not matter now. If we
have to use the font ic, then I will make it work.